micro-tours-mvp/backend/requirements.txt
fastapi
uvicorn[standard]
celery[redis]
redis
python-dotenv
moviepy
jinja2
requests
pydantic
pytest
pytest-cov
httpx
